requirements.txt
ImageHash==4.3.1
numpy==1.26.2
Pillow==10.1.0
pillow_heif==0.14.0
